Add parallelization flag to ray ground classifier
Ray ground classifier has both single thread and parallel
implementations. ray_ground_classifier and ray_ground_classifier_nodes
are coupled such that they have to be compiled with the same
parallelization flag. This is error prone in that two cmakefiles need to
be modified in order to turn off parallelization correctly.

This patch add a cmake macro that can be set on command line to turn off
parallelization for both modules.

Change-Id: Ida81cdc24ae1d22f699ed2cd79eeefec576fe384
Signed-off-by: Liyou Zhou <liyou.zhou@arm.com>
Issue-Id: SCM-1007