[#897] Fix misassociation bug in hungarian_assigner
Hungarian assigner was giving sub optimal association in certain
degenerate input cases. This commit changes the logic to use a max
weight for pairs with no set weights, instead of float max it now uses a
number much less so that the whole assignment problem can be solved as
if all the pairs have a valid weight and the get_assignment() takes care
of returning UNASSIGNED for pairs that did not originally have a set
weight.